video-agent/audio-agent/requirements.txt
openai>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
elevenlabs>=1.0.0
